# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.HandleMarkBookForSubscriptionParams:
    properties:
      subscription:
        type: boolean
    required:
    - subscription
    type: object
...
host: localhost:8080
info:
  contact: {}
  title: Pagesy
  version: "1.0"
paths:
...
  /books/{bookId}/subscriptions:
    patch:
      description: Mark book for subscription
      parameters:
      - description: book id
        in: path
        name: bookId
        required: true
        type: string
      - description: mark book for subscription body
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/models.HandleMarkBookForSubscriptionParams'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Mark book for subscription
      tags:
      - subscriptions
...
swagger: "2.0"